requirements.txt
# Web framework
Flask==3.0.3
Werkzeug==3.0.3

# PostgreSQL
psycopg2-binary==2.9.9

# Auth
bcrypt==4.1.3

# Payments
razorpay==1.4.1

# Supabase Storage
supabase==2.4.2

# HTTP requests
requests==2.31.0

# Environment
python-dotenv==1.0.1

# Production server
gunicorn==22.0.0

# CSRF protection
Flask-WTF==1.2.1

# Rate limiting
Flask-Limiter==3.7.0
